requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asyncpg
attrs
cachetools
certifi
charset-normalizer
dataclasses-json
distro
filetype
frozenlist
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httpx
httpx-sse
idna
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-postgres
langchain-text-splitters
langsmith
marshmallow
multidict
mypy_extensions
numpy
openai
orjson
packaging
pgvector
propcache
proto-plus
protobuf
psycopg
psycopg-binary
psycopg-pool
psycopg2-binary
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
jiter
pypdf
python-dotenv
PyYAML
regex
requests
requests-toolbelt
rsa
sniffio
SQLAlchemy
tenacity
tiktoken
tqdm
typing-inspect
typing-inspection
typing_extensions
urllib3
yarl
zstandard